first commit. after a plethora of problems, this was finally figured out. as per usual, i was making things much more harder than they needed to be. with a little direction from izzy and megan, i was able to figure out what i was doing incorrectly to get the images to show.
