+++ index.html
@@ -1,15 +1,15 @@
 <!DOCTYPE html>
 <html>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width" />
+    <title>Import Export Practice- FMS</title>
+    <link rel="stylesheet" href="./css/style.css" />
+  </head>
 
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width">
-    <title>Import Export Practice</title>
-</head>
-
-<body>
+  <body>
     <h1 class="shadow">Artists</h1>
     <div id="list"></div>
-</body>
-
+    <script type="module" src="./javascript/app.js"></script>
+  </body>
 </html>
