--- style.css
@@ -1,45 +1,0 @@
-body {
-    background-image: url("./background_colors.jpg");
-    background-repeat: no-repeat;
-    background-size: 100%;
-    margin: 0px;
-}
-
-h1 {
-    text-align: center;
-    background: #d8d8d8;
-    padding: 30px;
-    margin: 0px;
-}
-
-.shadow {
-    box-shadow: 5px 7px 5px 5px #00000057;
-}
-
-.profile {
-    display: inline-block;
-    background: #d8d8d8;
-    overflow: hidden;
-    margin: 20px;
-}
-
-.profile img {
-    width: 200px;
-    height: 200px;
-    margin: 20px;
-    border-radius: 100px;
-}
-
-.profile h2 {
-    text-align: center;
-}
-
-#list ul {
-    list-style: none;
-    display: flex;
-    justify-content: space-evenly;
-}
-
-#list li {
-    display: inline-block;
-}
